pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.84.0",
    "pytz",
    "python-dateutil",
    "beautifulsoup4==4.11.2",
    "bs4==0.0.1",
    "certifi==2022.12.7",
    "charset-normalizer==3.1.0",
    "idna==3.4",
    "requests==2.28.2",
    "soupsieve==2.4",
    "urllib3==1.26.15",
    "bilibili-api-python>=17.2.0",
    "httpx>=0.28.1",
    "yarl>=1.20.0",
    "pillow>=11.2.1",
    "pycryptodome>=3.23.0",
    "pycryptodomex>=3.23.0",
    "pyyaml>=6.0.2",
    "brotli~=1.1.0",
    "pyjwt>=2.10.1",
    "colorama~=0.4.6",
    "lxml~=5.4.0",
    "pyyaml~=6.0",
    "brotli~=1.1.0",
    "qrcode~=8.2",
    "APScheduler~=3.11.0",
    "qrcode_terminal~=0.8",
    "socksio>=1.0.0",
]
